# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: WhatsApp API Backend
  version: 1.0.0
  description: API management platform for integrating WhatsApp messaging into applications.
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /users:
    post:
      summary: Create a new user (Admin Only)
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: newuser@example.com
                password:
                  type: string
                  example: securepassword123
                name:
                  type: string
                  example: New User
                role:
                  type: string
                  enum: [USER, ADMINISTRATOR]
                  default: USER
              required:
                - email
                - password
                - name
      responses:
        201:
          description: User created successfully
        400:
          description: Invalid input
        403:
          description: Forbidden (Admin role required)
        409:
          description: User with email already exists
    get:
      summary: Get all users (Admin Only)
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
          description: Number of users per page
        - in: query
          name: sort
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
        - in: query
          name: sort_by
          schema:
            type: string
            default: createdAt
          description: Field to sort by
      responses:
        200:
          description: List of users
        403:
          description: Forbidden (Admin role required)
...
  /instances:
    post:
      summary: Create a new WhatsApp instance
      tags:
        - Instances
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Unique name for the instance
                  example: my-whatsapp-instance
                displayName:
                  type: string
                  description: Display name for the instance
                  example: My WhatsApp Business
                description:
                  type: string
                  description: Description of the instance
                  example: This is a test instance for development.
              required:
                - name
      responses:
        201:
          description: Instance created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      status:
                        type: string
                  meta:
                    type: object
        400:
          description: Invalid input
        403:
          description: Subscription limit reached
    get:
      summary: List all WhatsApp instances for the current user
      tags:
        - Instances
      security:
        - BearerAuth: []
      responses:
        200:
          description: List of instances
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        status:
                          type: string
                  meta:
                    type: object
...
  /stats:
    get:
      summary: Get message statistics for the current instance (API Key context)
      tags:
        - Instances
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
            example: 2024-01-01
          description: Start date for statistics (YYYY-MM-DD)
        - in: query
          name: endDate
          schema:
            type: string
            format: date
            example: 2024-01-31
          description: End date for statistics (YYYY-MM-DD)
      responses:
        200:
          description: Message statistics
        401:
          description: Unauthorized (invalid API key)
  /webhook:
    post:
      summary: Configure webhook (Integration API)
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: url
                  example: https://your-app-webhook.com/whatsapp-events
                events:
                  type: object
                  additionalProperties:
                    type: boolean
                  example: { "message.received": true, "instance.status": true }
                headers:
                  type: object
                  additionalProperties:
                    type: string
                  example: { "X-App-Token": "your-app-token" }
                isActive:
                  type: boolean
                  example: true
              required:
                - url
                - events
      responses:
        200:
          description: Webhook configured successfully
        400:
          description: Invalid input
    get:
      summary: Get webhook configuration (Integration API)
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      responses:
        200:
          description: Webhook configuration
        404:
          description: Webhook not found
    put:
      summary: Update webhook configuration (Integration API)
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: url
                  example: https://updated-app-webhook.com/whatsapp-events
                events:
                  type: object
                  additionalProperties:
                    type: boolean
                  example: { "message.received": true, "message.sent": true }
                headers:
                  type: object
                  additionalProperties:
                    type: string
                  example: { "X-App-Token": "updated-app-token" }
                isActive:
                  type: boolean
                  example: true
              required: []
      responses:
        200:
          description: Webhook updated successfully
        400:
          description: Invalid input
        404:
          description: Webhook not found
    delete:
      summary: Delete webhook configuration (Integration API)
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      responses:
        200:
          description: Webhook deleted successfully
        404:
          description: Webhook not found
...
  /webhook/stats:
    get:
      summary: Get webhook statistics (Integration API)
      tags:
        - Webhooks
      security:
        - BearerAuth: []
      responses:
        200:
          description: Webhook statistics
        404:
          description: Webhook not configured
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT